+++ docling/datamodel/service/__init__.py
@@ -26,6 +26,7 @@ from docling.datamodel.service.options import (
 )
 from docling.datamodel.service.requests import (
     AnyHttpSourceRequest,
+    AzureBlobSourceRequest,
     BaseChunkDocumentsRequest,
     BatchConvertSourcesRequest,
     BatchSourceRequestItem,
@@ -34,6 +35,8 @@ from docling.datamodel.service.requests import (
     ConvertSourcesRequest,
     FileSourceRequest,
     GenericChunkDocumentsRequest,
+    GoogleCloudStorageSourceRequest,
+    GoogleDriveSourceRequest,
     HttpSourceRequest,
     S3SourceRequest,
     SourceRequestItem,
@@ -69,9 +72,21 @@ from docling.datamodel.service.responses import (
     WebsocketMessage,
     ZipArchiveResult,
 )
-from docling.datamodel.service.sources import FileSource, HttpSource, S3Coordinates
+from docling.datamodel.service.sources import (
+    AzureBlobCoordinates,
+    FileSource,
+    GoogleCloudStorageCoordinates,
+    GoogleCloudStorageServiceAccountInfo,
+    GoogleDriveCoordinates,
+    GoogleDriveCredentials,
+    HttpSource,
+    S3Coordinates,
+)
 from docling.datamodel.service.targets import (
     ChunkTarget,
+    AzureBlobTarget,
+    GoogleCloudStorageTarget,
+    GoogleDriveTarget,
     InBodyTarget,
     PresignedUrlTarget,
     PutTarget,
@@ -84,6 +99,9 @@ from docling.datamodel.service.tasks import TaskProcessingMeta, TaskType
 __all__ = [
     "AnyHttpSourceRequest",
     "ArtifactRef",
+    "AzureBlobCoordinates",
+    "AzureBlobSourceRequest",
+    "AzureBlobTarget",
     "BaseChunkDocumentsRequest",
     "BaseChunkerOptions",
     "BaseProgress",
@@ -114,6 +132,14 @@ __all__ = [
     "FileSource",
     "FileSourceRequest",
     "GenericChunkDocumentsRequest",
+    "GoogleCloudStorageCoordinates",
+    "GoogleCloudStorageServiceAccountInfo",
+    "GoogleCloudStorageSourceRequest",
+    "GoogleCloudStorageTarget",
+    "GoogleDriveCoordinates",
+    "GoogleDriveCredentials",
+    "GoogleDriveSourceRequest",
+    "GoogleDriveTarget",
     "HealthCheckResponse",
     "HierarchicalChunkerOptions",
     "HttpSource",